# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/transactions": {
      "post": {
        "operationId": "TransactionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new transaction (part or toner)",
        "tags": [
          "transactions"
        ]
      },
      "get": {
        "operationId": "TransactionsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all transactions (part and toner)",
        "tags": [
          "transactions"
        ]
      }
    },
    "/transactions/recent": {
      "get": {
        "operationId": "TransactionsController_getRecent",
        "parameters": [
          {
            "name": "hours",
            "required": false,
            "in": "query",
            "schema": {
              "example": 12,
              "type": "number"
            }
          },
          {
            "name": "types",
            "required": false,
            "in": "query",
            "description": "Comma-separated transaction types (e.g. claim,collect,request)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get recent transactions by hours and types",
        "tags": [
          "transactions"
        ]
      }
    },
    "/transactions/collected": {
      "get": {
        "operationId": "TransactionsController_getRecentCollections",
        "parameters": [
          {
            "name": "hours",
            "required": false,
            "in": "query",
            "schema": {
              "example": 12,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get recent collected transactions by hours",
        "tags": [
          "transactions"
        ]
      }
    },
    "/transactions/{id}": {
      "get": {
        "operationId": "TransactionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Transaction UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a transaction by ID (part or toner)",
        "tags": [
          "transactions"
        ]
      }
    },
    "/toners": {
      "post": {
        "operationId": "TonersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTonerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Toners"
        ]
      },
      "get": {
        "operationId": "TonersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Toners"
        ]
      }
    },
    "/toners/{id}": {
      "get": {
        "operationId": "TonersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Toners"
        ]
      },
      "patch": {
        "operationId": "TonersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTonerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Toners"
        ]
      },
      "delete": {
        "operationId": "TonersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Toners"
        ]
      }
    },
    "/toners/{id}/claim": {
      "patch": {
        "operationId": "TonersController_claimToner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Toners"
        ]
      }
    },
    "/toners/{id}/collect": {
      "patch": {
        "operationId": "TonersController_collectToner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Toners"
        ]
      }
    }
  },
  "info": {
    "title": "Parts Management API",
    "description": "Endpoints to manage parts, claims, and users",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "partId": {
            "type": "string",
            "format": "uuid"
          },
          "tonerId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "CLAIM",
              "REQUEST",
              "RETURN",
              "COLLECT",
              "ADD"
            ]
          },
          "quantityDelta": {
            "type": "number",
            "minimum": 1
          },
          "details": {
            "type": "string",
            "description": "Optional notes/details"
          }
        },
        "required": [
          "userId",
          "type",
          "quantityDelta"
        ]
      },
      "Part": {
        "type": "object",
        "properties": {}
      },
      "Toner": {
        "type": "object",
        "properties": {}
      },
      "User": {
        "type": "object",
        "properties": {}
      },
      "TransactionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "part": {
            "$ref": "#/components/schemas/Part"
          },
          "toner": {
            "$ref": "#/components/schemas/Toner"
          },
          "type": {
            "type": "string",
            "enum": [
              "CLAIM",
              "REQUEST",
              "RETURN",
              "COLLECT",
              "ADD"
            ]
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "quantityDelta": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "type",
          "user",
          "quantityDelta",
          "createdAt"
        ]
      },
      "CreateTonerDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateTonerDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}